pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "chardet>=5.2.0",
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "patool>=4.0.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "python-dateutil>=2.9.0.post0",
    "python-magic>=0.4.27",
    "pyyaml>=6.0.2",
    "rarfile>=4.2",
    "werkzeug>=3.1.3",
    "flask-wtf>=1.2.2",
    "wtforms>=3.2.1",
    "supabase>=2.20.0",
    "python-dotenv>=1.1.1",
    "alembic>=1.16.5",
    "google-cloud-storage>=3.4.1",
    "requests>=2.32.5",
    "psutil>=7.1.1",
]
